Direct3D Rendering
I have finally implemented a very very basic version of Direct3D rendering into the program by merging the two frameworks from waynes code. This has taken two days of work to fully understand what is going on where and when but it has been a good learning experience on the basics of using the DirectX API.

This can and probably will be revised over the comming weeks but at this moment in time it works, next thing to do is to implement scene graphs and nodes so that I can actually render objects in the program.